Rename getunitsdata.{cpp,h} to unitsdata.{cpp,h}
