+++ icu4c/source/i18n/i18n_uwp.vcxproj
@@ -488,7 +488,7 @@
     <ClCompile Include="unitconverter.cpp" />
     <ClCompile Include="complexunitsconverter.cpp" />
     <ClCompile Include="unitsrouter.cpp" />
-    <ClCompile Include="getunitsdata.cpp" />
+    <ClCompile Include="unitsdata.cpp" />
     <ClCompile Include="unum.cpp" />
     <ClCompile Include="unumsys.cpp" />
     <ClCompile Include="upluralrules.cpp" />
@@ -726,7 +726,7 @@
     <ClInclude Include="unitconverter.h" />
     <ClInclude Include="complexunitsconverter.h" />
     <ClInclude Include="unitsrouter.h" />
-    <ClInclude Include="getunitsdata.h" />
+    <ClInclude Include="unitsdata.h" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="i18n.rc" />
