+++ icu4c/source/i18n/unitconverter.cpp
@@ -9,7 +9,6 @@
 
 #include "charstr.h"
 #include "double-conversion.h"
-#include "getunitsdata.h"
 #include "measunit_impl.h"
 #include "resource.h"
 #include "uassert.h"
