+++ icu4c/source/i18n/unitsdata.cpp
@@ -9,9 +9,9 @@
 
 #include "cmemory.h"
 #include "cstring.h"
-#include "getunitsdata.h"
 #include "number_decimalquantity.h"
 #include "resource.h"
+#include "unitsdata.h"
 #include "uresimp.h"
 
 U_NAMESPACE_BEGIN
