+++ icu4c/source/i18n/unitsrouter.cpp
@@ -9,7 +9,6 @@
 
 #include "cmemory.h"
 #include "cstring.h"
-#include "getunitsdata.h"
 #include "number_decimalquantity.h"
 #include "resource.h"
 #include "unitsrouter.h"
